console/pyproject.toml
[project]
name = "console"
version = "0.1.0"
description = "Interactive console for AI-DB and AI-Frontend with natural language interface"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ai-db @ file:///${PROJECT_ROOT}/ai-db",
    "ai-frontend @ file:///${PROJECT_ROOT}/ai-frontend", 
    "git-layer @ file:///${PROJECT_ROOT}/git-layer",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "dependency-injector>=4.0.0",
    "click>=8.0.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
console = "console.main:main"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
target-version = "py313"
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM", "RUF"]